+++ Muse_performance/app/src/main/java/muse/muse_performance/MidiFileWriter.java
@@ -6,9 +6,7 @@ import android.content.Context;
 
 public class MidiFileWriter
 {
-	private String FILE_ID = ".mid";
-
-	//　MThd　データ
+	//　MThd データ
 	private byte[] HEADER_TAG = {0x4d,0x54,0x68,0x64};
 	private byte[] HEADER_DATA_LEN = {0x00,0x00,0x00,0x06};
 	private byte[] MIDI_FORMAT = {0x00,0x01};
@@ -18,569 +16,16 @@ public class MidiFileWriter
 	private byte[] TRACK_END = {0x00,(byte)0xff,0x2f,0x00};
 
 	private FileOutputStream fos = null;
+	private String FILE_ID = ".mid";
 	private Context context = null;
 	private ArrayList<TrackData> TrackDataList = new ArrayList<TrackData>();
 	private int quarterNoteDeltaTime;
 
 	// コンストラクタ
-	public MidiFileWriter(Context contxt)
+	public MidiFileWriter(Context context)
 	{
-		this.context = contxt;
+		this.context = context;
 	}
-
-	//オリジナル１曲目
-	protected void Song1Melody( byte ch, byte vel )
-	{
-		try
-		{
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_010), (byte)MidiFileWriter.NoteTone.G5, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.FF5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_020), (byte)MidiFileWriter.NoteTone.FF5, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_020), (byte)MidiFileWriter.NoteTone.G5, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_025), (byte)MidiFileWriter.NoteTone.D5, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)MidiFileWriter.NoteTone.B4, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_010), (byte)MidiFileWriter.NoteTone.G4, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_010), (byte)MidiFileWriter.NoteTone.G4, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.A4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_025), (byte)MidiFileWriter.NoteTone.A4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)MidiFileWriter.NoteTone.B4, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_010), (byte)MidiFileWriter.NoteTone.D5, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.E5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_025), (byte)MidiFileWriter.NoteTone.E5, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)MidiFileWriter.NoteTone.D5, (byte)0x00);
-
-			this.closeTrackData();
-		}
-		catch( Exception e )
-		{
-			e.printStackTrace();
-		}
-	}
-
-	// 打楽器
-	protected void Song1PercussionA( byte ch, byte vel )
-	{
-		try
-		{
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)54, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)54, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)54, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)54, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)54, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)54, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)54, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)54, (byte)0x00);
-
-			this.closeTrackData();
-		}
-		catch( Exception e )
-		{
-			e.printStackTrace();
-		}
-	}
-
-	protected void Song1PercussionB( byte ch, byte vel )
-	{
-		try
-		{
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)54, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)54, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)54, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)54, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)54, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)54, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)54, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)35, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)35, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_040), (byte)54, (byte)0x00);
-
-			this.closeTrackData();
-		}
-		catch( Exception e )
-		{
-			e.printStackTrace();
-		}
-	}
-
-	protected void Song1Percuss( int no, byte ch, byte vel )
-	{
-		switch( no )
-		{
-			case 0:
-				this.Song1PercussionA( ch, vel);
-				break;
-
-			case 1:
-				this.Song1PercussionB( ch, vel );
-				break;
-
-			default:
-				this.Song1PercussionA( ch, vel);
-				break;
-		}
-	}
-
-	protected void Song1ArrangeA( byte ch, byte vel )
-	{
-		try
-		{
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.E3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.E3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.E3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.E3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_020), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G2, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.G2, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G2, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.G2, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G2, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.G3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.FF3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.G3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.FF3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_020), (byte)MidiFileWriter.NoteTone.G3, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.E3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.E3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.E3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.E3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_020), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G2, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.G2, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G2, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.G2, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G2, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.G3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.FF3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.G3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.FF3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_020), (byte)MidiFileWriter.NoteTone.G3, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.E3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.E3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.E3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.E3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_020), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G2, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.G2, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G2, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.G2, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G2, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.G3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.FF3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.G3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.FF3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_020), (byte)MidiFileWriter.NoteTone.G3, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.E3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.E3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.E3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.C3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.E3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_020), (byte)MidiFileWriter.NoteTone.C3, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G2, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.G2, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G2, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.G2, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G2, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.G3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.FF3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.G3, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D3, vel);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.FF3, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_020), (byte)MidiFileWriter.NoteTone.G3, (byte)0x00);
-
-			this.closeTrackData();
-		}
-		catch( Exception e )
-		{
-			e.printStackTrace();
-		}
-	}
-
-	protected void Song1ArrangeB( byte ch, byte vel )
-	{
-		try
-		{
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.A4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.A4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.A4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.A4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.A4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.A4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.A4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.A4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.G4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.D5, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-			this.addNoteOn( ch, 0, (byte)MidiFileWriter.NoteTone.B4, vel);
-			this.addNoteOn( ch, this.getNoteDeltaTime(MidiFileWriter.NoteTime.Note_080), (byte)MidiFileWriter.NoteTone.C4, (byte)0x00);
-
-			this.closeTrackData();
-		}
-		catch( Exception e )
-		{
-			e.printStackTrace();
-		}
-	}
-
-	// 2曲目
 
 	// midiファイル作成
 	public boolean CreateMidiFile( String fileName, int trackNo, int dTime )
@@ -699,7 +144,6 @@ public class MidiFileWriter
 		addTrackData(data, dataLen);
 	}
 
-
 	// Trackデータクローズ
 	public void closeTrackData()
 	{
@@ -806,147 +250,1087 @@ public class MidiFileWriter
 		return dltTime;
 	}
 
+	// 曲譜面選択
+	public void SelectSong( int songNo, int dNo, int[] vel )
+	{
+		switch( songNo )
+		{
+			case 0:
+				this.Song1ArrangeA((byte) 0x00, (byte) vel[0]);
+				this.Song1ArrangeB((byte) 0x01, (byte) vel[1]);
+				this.Song1Melody((byte) 0x02, (byte) vel[2]);
+				this.Song1Percuss( dNo, (byte) vel[3]);
+				break;
+
+			case 1:
+				this.Song2ArrangeA((byte) 0x00, (byte) vel[0]);
+				this.Song2ArrangeB((byte) 0x01, (byte) vel[1]);
+				this.Song2Melody((byte) 0x02, (byte) vel[2]);
+				this.Song2Percuss( dNo, (byte) vel[3]);
+				break;
+
+			default:
+				this.Song1ArrangeA((byte) 0x00, (byte) vel[0]);
+				this.Song1ArrangeB((byte) 0x01, (byte) vel[1]);
+				this.Song1Melody((byte) 0x02, (byte) vel[2]);
+				this.Song1PercussionA((byte) 0x09, (byte) vel[3]);
+				break;
+		}
+	}
+
+	/* オリジナル1 */
+	//主旋律
+	protected void Song1Melody( byte ch, byte vel )
+	{
+		try
+		{
+			this.addNoteOn( ch, 0, (byte)NoteTone.G5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_010), (byte)NoteTone.G5, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_020), (byte)NoteTone.FF5, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_020), (byte)NoteTone.G5, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_025), (byte)NoteTone.D5, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.B4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.G4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_010), (byte)NoteTone.G4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.G4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_010), (byte)NoteTone.G4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.A4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_025), (byte)NoteTone.A4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.B4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_010), (byte)NoteTone.D5, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.E5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_025), (byte)NoteTone.E5, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.D5, (byte)0x00);
+
+			this.closeTrackData();
+		}
+		catch( Exception e )
+		{
+			e.printStackTrace();
+		}
+	}
+
+	// ドラムA
+	protected void Song1PercussionA( byte ch, byte vel )
+	{
+		try
+		{
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.closeTrackData();
+		}
+		catch( Exception e )
+		{
+			e.printStackTrace();
+		}
+	}
+
+	// ドラムB
+	protected void Song1PercussionB( byte ch, byte vel )
+	{
+		try
+		{
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.closeTrackData();
+		}
+		catch( Exception e )
+		{
+			e.printStackTrace();
+		}
+	}
+
+	// ドラム管理
+	protected void Song1Percuss( int no, byte vel )
+	{
+		switch( no )
+		{
+			case 0:
+				this.Song1PercussionA( (byte) 0x09, vel);
+				break;
+
+			case 1:
+				this.Song1PercussionB( (byte) 0x09, vel );
+				break;
+
+			default:
+				this.Song1PercussionA( (byte) 0x09, vel);
+				break;
+		}
+	}
+
+	// 伴奏A
+	protected void Song1ArrangeA( byte ch, byte vel )
+	{
+		try
+		{
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.E3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.E3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.E3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.E3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_020), (byte)NoteTone.C3, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.G2, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.G2, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G2, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.G2, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G2, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.G3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.G3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_020), (byte)NoteTone.G3, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.E3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.E3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.E3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.E3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_020), (byte)NoteTone.C3, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.G2, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.G2, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G2, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.G2, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G2, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.G3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.G3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_020), (byte)NoteTone.G3, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.E3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.E3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.E3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.E3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_020), (byte)NoteTone.C3, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.G2, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.G2, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G2, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.G2, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G2, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.G3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.G3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_020), (byte)NoteTone.G3, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.E3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.E3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.E3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.C3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.E3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_020), (byte)NoteTone.C3, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.G2, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.G2, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G2, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.G2, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G2, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.G3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.G3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_020), (byte)NoteTone.G3, (byte)0x00);
+
+			this.closeTrackData();
+		}
+		catch( Exception e )
+		{
+			e.printStackTrace();
+		}
+	}
+
+	// 伴奏B
+	protected void Song1ArrangeB( byte ch, byte vel )
+	{
+		try
+		{
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.B4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+
+			this.closeTrackData();
+		}
+		catch( Exception e )
+		{
+			e.printStackTrace();
+		}
+	}
+
+	/* オリジナル2 */
+	//主旋律
+	protected void Song2Melody( byte ch, byte vel )
+	{
+		try
+		{
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_010), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.A5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_025), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.E5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_010), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.G5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_010), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.D4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_010), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.A5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_025), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.E5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_010), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.A5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_020), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_020), (byte)NoteTone.C4, (byte)0x00);
+
+			this.closeTrackData();
+		}
+		catch( Exception e )
+		{
+			e.printStackTrace();
+		}
+	}
+
+	// ドラムA
+	protected void Song2PercussionA( byte ch, byte vel )
+	{
+		try
+		{
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+
+			this.closeTrackData();
+		}
+		catch( Exception e )
+		{
+			e.printStackTrace();
+		}
+	}
+
+	// ドラムB
+	protected void Song2PercussionB( byte ch, byte vel )
+	{
+		try
+		{
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)35, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_160), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)35, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)40, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)54, (byte)0x00);
+
+			this.closeTrackData();
+		}
+		catch( Exception e )
+		{
+			e.printStackTrace();
+		}
+	}
+
+	// ドラム管理
+	protected void Song2Percuss( int no, byte vel )
+	{
+		switch( no )
+		{
+			case 0:
+				this.Song2PercussionA( (byte) 0x09, vel);
+				break;
+
+			case 1:
+				this.Song2PercussionB( (byte) 0x09, vel );
+				break;
+
+			default:
+				this.Song2PercussionA( (byte) 0x09, vel);
+				break;
+		}
+	}
+
+	// 伴奏A
+	protected void Song2ArrangeA( byte ch, byte vel )
+	{
+		try
+		{
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			// 小節またぐ　080
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C3, (byte)0x00);
+			//             080
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C3, (byte)0x00);this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			// 小節またぐ　080
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C3, (byte)0x00);
+			//             080
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C3, (byte)0x00);this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			// 小節またぐ　080
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C3, (byte)0x00);
+			//             080
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C3, (byte)0x00);this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			// 小節またぐ　080
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C3, (byte)0x00);
+			//             080
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C3, (byte)0x00);this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C3, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);this.addNoteOn( ch, 0, (byte)NoteTone.FF3, vel);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D3, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_080), (byte)NoteTone.C3, (byte)0x00);
+
+			this.closeTrackData();
+		}
+		catch( Exception e )
+		{
+			e.printStackTrace();
+		}
+	}
+
+	// 伴奏B
+	protected void Song2ArrangeB( byte ch, byte vel )
+	{
+		try
+		{
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.CC5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_010), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.CC5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_010), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.CC5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.A4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_010), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.FF5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.G4, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C4, (byte)0x00);
+			this.addNoteOn( ch, 0, (byte)NoteTone.D5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_040), (byte)NoteTone.C4, (byte)0x00);
+
+			this.addNoteOn( ch, 0, (byte)NoteTone.E5, vel);
+			this.addNoteOn( ch, this.getNoteDeltaTime(NoteTime.Note_010), (byte)NoteTone.C4, (byte)0x00);
+
+
+
+			this.closeTrackData();
+		}
+		catch( Exception e )
+		{
+			e.printStackTrace();
+		}
+	}
+
 	/* 定数 */
 	// Trackデータ
 	protected class TrackData
 	{
 		protected int length = 0;
 		protected byte[] data;
-	}
-
-	// NOTE LEN
-	public class NoteTime
-	{
-		public final static double Note_010 = 4;    // 全音符
-		public final static double Note_025 = 3;    // 付点2分音符
-		public final static double Note_020 = 2;    // 2分音符
-		public final static double Note_045 = 1.5;  // 付点4分音符
-		public final static double Note_040 = 1;    // 4分音符
-		public final static double Note_085 = 0.75; // 付点8分音符
-		public final static double Note_080 = 0.5;  // 8分音符
-		public final static double Note_160 = 0.25; // 16分音符
-	}
-
-	// NOTE TONE
-	public class NoteTone
-	{
-		public final static int C0 = 12;
-		public final static int CC0 = 13;
-		public final static int D0 = 14;
-		public final static int DD0 = 15;
-		public final static int E0 = 16;
-		public final static int F0 = 17;
-		public final static int FF0 = 18;
-		public final static int G0 = 19;
-		public final static int GG0 = 20;
-		public final static int A0 = 21;
-		public final static int AA0 = 22;
-		public final static int B0 = 23;
-
-		public final static int C1 = 24;
-		public final static int CC1 = 25;
-		public final static int D1 = 26;
-		public final static int DD1 = 27;
-		public final static int E1 = 28;
-		public final static int F1 = 29;
-		public final static int FF1 = 30;
-		public final static int G1 = 31;
-		public final static int GG1 = 32;
-		public final static int A1 = 33;
-		public final static int AA1 = 34;
-		public final static int B1 = 35;
-		public final static int C2 = 36;
-		public final static int CC2 = 37;
-		public final static int D2 = 38;
-		public final static int DD2 = 39;
-		public final static int E2 = 40;
-		public final static int F2 = 41;
-		public final static int FF2 = 42;
-		public final static int G2 = 43;
-		public final static int GG2 = 44;
-		public final static int A2 = 45;
-		public final static int AA2 = 46;
-		public final static int B2 = 47;
-		public final static int C3 = 48;
-		public final static int CC3 = 49;
-		public final static int D3 = 50;
-		public final static int DD3 = 51;
-		public final static int E3 = 52;
-		public final static int F3 = 53;
-		public final static int FF3 = 54;
-		public final static int G3 = 55;
-		public final static int GG3 = 56;
-		public final static int A3 = 57;
-		public final static int AA3 = 58;
-		public final static int B3 = 59;
-		public final static int C4 = 60;
-		public final static int CC4 = 61;
-		public final static int D4 = 62;
-		public final static int DD4 = 63;
-		public final static int E4 = 64;
-		public final static int F4 = 65;
-		public final static int FF4 = 66;
-		public final static int G4 = 67;
-		public final static int GG4 = 68;
-		public final static int A4 = 69;
-		public final static int AA4 = 70;
-		public final static int B4 = 71;
-
-		public final static int C5 = 72;
-		public final static int CC5 = 73;
-		public final static int D5 = 74;
-		public final static int DD5 = 75;
-		public final static int E5 = 76;
-		public final static int F5 = 77;
-		public final static int FF5 = 78;
-		public final static int G5 = 79;
-		public final static int GG5 = 80;
-		public final static int A5 = 81;
-		public final static int AA5 = 82;
-		public final static int B5 = 83;
-		public final static int C6 = 84;
-		public final static int CC6 = 85;
-		public final static int D6 = 86;
-		public final static int DD6 = 87;
-		public final static int E6 = 88;
-		public final static int F6 = 89;
-		public final static int FF6 = 90;
-		public final static int G6 = 91;
-		public final static int GG6 = 92;
-		public final static int A6 = 93;
-		public final static int AA6 = 94;
-		public final static int B6 = 95;
-		public final static int C7 = 96;
-		public final static int CC7 = 97;
-		public final static int D7 = 98;
-		public final static int DD7 = 99;
-		public final static int E7 = 100;
-		public final static int F7 = 101;
-		public final static int FF7 = 102;
-		public final static int G7 = 103;
-		public final static int GG7 = 104;
-		public final static int A7 = 105;
-		public final static int AA7 = 106;
-		public final static int B7 = 107;
-		public final static int C8 = 108;
-		public final static int CC8 = 109;
-		public final static int D8 = 110;
-		public final static int DD8 = 111;
-		public final static int E8 = 112;
-		public final static int F8 = 113;
-		public final static int FF8 = 114;
-		public final static int G8 = 115;
-		public final static int GG8 = 116;
-		public final static int A8 = 117;
-		public final static int AA8 = 118;
-		public final static int B8 = 119;
-		public final static int C9 = 120;
-		public final static int CC9 = 121;
-		public final static int D9 = 122;
-		public final static int DD9 = 123;
-		public final static int E9 = 124;
-		public final static int F9 = 125;
-		public final static int FF9 = 126;
-		public final static int G9 = 127;
 	}
 }
