+++ Muse_performance/app/src/main/java/muse/muse_performance/NoteTime.java
@@ -1,0 +1,16 @@
+package muse.muse_performance;
+
+/**
+ * Created by Ryo on 2016/09/24.
+ */
+public class NoteTime
+{
+	public final static double Note_010 = 4;    // 全音符
+	public final static double Note_025 = 3;    // 付点2分音符
+	public final static double Note_020 = 2;    // 2分音符
+	public final static double Note_045 = 1.5;  // 付点4分音符
+	public final static double Note_040 = 1;    // 4分音符
+	public final static double Note_085 = 0.75; // 付点8分音符
+	public final static double Note_080 = 0.5;  // 8分音符
+	public final static double Note_160 = 0.25; // 16分音符
+}
