+++ Muse_performance/app/src/main/java/muse/muse_performance/NoteTone.java
@@ -1,0 +1,128 @@
+package muse.muse_performance;
+
+/**
+ * Created by ryo on 2016/09/24.
+ */
+public class NoteTone
+{
+	private NoteTone(){}
+
+	public final static int C0 = 12;
+	public final static int CC0 = 13;
+	public final static int D0 = 14;
+	public final static int DD0 = 15;
+	public final static int E0 = 16;
+	public final static int F0 = 17;
+	public final static int FF0 = 18;
+	public final static int G0 = 19;
+	public final static int GG0 = 20;
+	public final static int A0 = 21;
+	public final static int AA0 = 22;
+	public final static int B0 = 23;
+
+	public final static int C1 = 24;
+	public final static int CC1 = 25;
+	public final static int D1 = 26;
+	public final static int DD1 = 27;
+	public final static int E1 = 28;
+	public final static int F1 = 29;
+	public final static int FF1 = 30;
+	public final static int G1 = 31;
+	public final static int GG1 = 32;
+	public final static int A1 = 33;
+	public final static int AA1 = 34;
+	public final static int B1 = 35;
+	public final static int C2 = 36;
+	public final static int CC2 = 37;
+	public final static int D2 = 38;
+	public final static int DD2 = 39;
+	public final static int E2 = 40;
+	public final static int F2 = 41;
+	public final static int FF2 = 42;
+	public final static int G2 = 43;
+	public final static int GG2 = 44;
+	public final static int A2 = 45;
+	public final static int AA2 = 46;
+	public final static int B2 = 47;
+	public final static int C3 = 48;
+	public final static int CC3 = 49;
+	public final static int D3 = 50;
+	public final static int DD3 = 51;
+	public final static int E3 = 52;
+	public final static int F3 = 53;
+	public final static int FF3 = 54;
+	public final static int G3 = 55;
+	public final static int GG3 = 56;
+	public final static int A3 = 57;
+	public final static int AA3 = 58;
+	public final static int B3 = 59;
+	public final static int C4 = 60;
+	public final static int CC4 = 61;
+	public final static int D4 = 62;
+	public final static int DD4 = 63;
+	public final static int E4 = 64;
+	public final static int F4 = 65;
+	public final static int FF4 = 66;
+	public final static int G4 = 67;
+	public final static int GG4 = 68;
+	public final static int A4 = 69;
+	public final static int AA4 = 70;
+	public final static int B4 = 71;
+
+	public final static int C5 = 72;
+	public final static int CC5 = 73;
+	public final static int D5 = 74;
+	public final static int DD5 = 75;
+	public final static int E5 = 76;
+	public final static int F5 = 77;
+	public final static int FF5 = 78;
+	public final static int G5 = 79;
+	public final static int GG5 = 80;
+	public final static int A5 = 81;
+	public final static int AA5 = 82;
+	public final static int B5 = 83;
+	public final static int C6 = 84;
+	public final static int CC6 = 85;
+	public final static int D6 = 86;
+	public final static int DD6 = 87;
+	public final static int E6 = 88;
+	public final static int F6 = 89;
+	public final static int FF6 = 90;
+	public final static int G6 = 91;
+	public final static int GG6 = 92;
+	public final static int A6 = 93;
+	public final static int AA6 = 94;
+	public final static int B6 = 95;
+	public final static int C7 = 96;
+	public final static int CC7 = 97;
+	public final static int D7 = 98;
+	public final static int DD7 = 99;
+	public final static int E7 = 100;
+	public final static int F7 = 101;
+	public final static int FF7 = 102;
+	public final static int G7 = 103;
+	public final static int GG7 = 104;
+	public final static int A7 = 105;
+	public final static int AA7 = 106;
+	public final static int B7 = 107;
+	public final static int C8 = 108;
+	public final static int CC8 = 109;
+	public final static int D8 = 110;
+	public final static int DD8 = 111;
+	public final static int E8 = 112;
+	public final static int F8 = 113;
+	public final static int FF8 = 114;
+	public final static int G8 = 115;
+	public final static int GG8 = 116;
+	public final static int A8 = 117;
+	public final static int AA8 = 118;
+	public final static int B8 = 119;
+	public final static int C9 = 120;
+	public final static int CC9 = 121;
+	public final static int D9 = 122;
+	public final static int DD9 = 123;
+	public final static int E9 = 124;
+	public final static int F9 = 125;
+	public final static int FF9 = 126;
+	public final static int G9 = 127;
+}
